+++ src/eca/llm_api.clj
@@ -53,9 +53,9 @@
             (when (llm-util/provider-api-key "anthropic" (get-in db [:auth "anthropic"]) config)
               [:api-key-found "anthropic/claude-sonnet-4.5"])
             (when (llm-util/provider-api-key "openai" (get-in db [:auth "openai"]) config)
-              [:api-key-found "openai/gpt-5.1"])
+              [:api-key-found "openai/gpt-5.2"])
             (when (get-in db [:auth "github-copilot" :api-key])
-              [:api-key-found "github-copilot/gpt-5.1"])
+              [:api-key-found "github-copilot/gpt-5.2"])
             (when-let [ollama-model (first (filter #(string/starts-with? % config/ollama-model-prefix) (keys (:models db))))]
               [:ollama-running ollama-model])
             [:default "anthropic/claude-sonnet-4.5"])]